remove optimize Op, replace with compact_buffers function
