+++ .github/workflows/fuzz.yml
@@ -25,6 +25,8 @@ jobs:
           submodules: "recursive"
       - uses: ./.github/actions/setup-rust
         with:
+          toolchain: nightly
+        with:
           rust-version: nightly
       - name: Install cargo fuzz
         run: cargo install --locked cargo-fuzz
@@ -40,7 +42,7 @@ jobs:
           AWS_ENDPOINT_URL: 'https://01e9655179bbec953276890b183039bc.r2.cloudflarestorage.com'
       - name: Run fuzzing target
         id: fuzz
-        run: RUST_BACKTRACE=1 cargo fuzz run file_io -j "$(nproc)" -- -max_total_time=7200
+        run: RUST_BACKTRACE=1 cargo +nightly fuzz run file_io -j "$(nproc)" -- -max_total_time=7200
         continue-on-error: true
       - name: Archive crash artifacts
         uses: actions/upload-artifact@v4
