+++ encodings/runend/src/ops.rs
@@ -1,12 +1,10 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
-use vortex_array::arrays::{ConstantArray, PrimitiveArray};
-use vortex_array::vtable::{OperationsVTable, ValidityHelper};
-use vortex_array::{Array, ArrayRef, IntoArray, ToCanonical};
-use vortex_buffer::Buffer;
-use vortex_dtype::{NativePType, match_each_native_ptype};
-use vortex_error::{VortexExpect, VortexResult};
+use vortex_array::arrays::ConstantArray;
+use vortex_array::vtable::OperationsVTable;
+use vortex_array::{Array, ArrayRef, IntoArray};
+use vortex_error::VortexResult;
 use vortex_scalar::Scalar;
 
 use crate::{RunEndArray, RunEndVTable};
@@ -47,29 +45,6 @@ impl OperationsVTable<RunEndVTable> for RunEndVTable {
     fn scalar_at(array: &RunEndArray, index: usize) -> VortexResult<Scalar> {
         array.values().scalar_at(array.find_physical_index(index)?)
     }
-
-    fn optimize(array: &RunEndArray) -> VortexResult<RunEndArray> {
-        // Only optimize sliced RunEndArray
-        if array.offset() == 0 {
-            return Ok(array.clone());
-        }
-
-        // Renumber the ends to account for the offset.
-        let ends = array.ends().to_primitive()?;
-        let new_ends = match_each_native_ptype!(ends.ptype(), |P| {
-            let buffer = renumber_ends(ends.as_slice::<P>(), array.offset());
-            PrimitiveArray::new(buffer, ends.validity().clone()).into_array()
-        });
-
-        // New values, sliced properly
-        RunEndArray::with_offset_and_length(new_ends, array.values().clone(), 0, array.len())
-    }
-}
-
-/// Subtract the offset out from all of the ends.
-fn renumber_ends<I: NativePType>(ends: &[I], offset: usize) -> Buffer<I> {
-    let offset = I::from(offset).vortex_expect("offset must fit into index type");
-    ends.iter().map(|&end| end - offset).collect()
 }
 
 #[cfg(test)]
@@ -79,7 +54,7 @@ mod tests {
     use vortex_buffer::buffer;
     use vortex_dtype::{DType, Nullability, PType};
 
-    use crate::{RunEndArray, RunEndVTable};
+    use crate::RunEndArray;
 
     #[test]
     fn slice_array() {
@@ -180,31 +155,5 @@ mod tests {
         .scalar_at(11)
         .unwrap();
         assert_eq!(scalar, 5.into());
-    }
-
-    #[test]
-    fn slice_and_optimize() {
-        // Encodes as
-        //   ends:   [4, 8]
-        //   values: [1, 2]
-        let original = RunEndArray::encode(
-            buffer![1u32, 1u32, 1u32, 1u32, 2u32, 2u32, 2u32, 2u32].into_array(),
-        )
-        .unwrap();
-
-        // After slicing:
-        //   ends:   [4, 8]
-        //   values: [1, 2]
-        let sliced = original.slice(2, 5).unwrap();
-
-        // After optimizing:
-        //  ends:   [0, 4]
-        //  values: [1, 2]
-        let optimized = sliced.optimize().unwrap();
-
-        assert_eq!(optimized.as_::<RunEndVTable>().offset(), 0);
-
-        let values = optimized.to_primitive().unwrap();
-        assert_eq!(values.as_slice::<u32>(), &[1u32, 1u32, 2u32]);
     }
 }
