+++ vortex-array/src/array/mod.rs
@@ -63,11 +63,6 @@ pub trait Array: 'static + private::Sealed + Send + Sync + Debug + ArrayVisitor 
     /// Fetch the scalar at the given index.
     fn scalar_at(&self, index: usize) -> VortexResult<Scalar>;
 
-    /// Return an optimized version of the same array.
-    ///
-    /// See [`OperationsVTable::optimize`] for more details.
-    fn optimize(&self) -> VortexResult<ArrayRef>;
-
     /// Returns whether the array is of the given encoding.
     fn is_encoding(&self, encoding: EncodingId) -> bool {
         self.encoding_id() == encoding
@@ -187,10 +182,6 @@ impl Array for Arc<dyn Array> {
 
     fn scalar_at(&self, index: usize) -> VortexResult<Scalar> {
         self.as_ref().scalar_at(index)
-    }
-
-    fn optimize(&self) -> VortexResult<ArrayRef> {
-        self.as_ref().optimize()
     }
 
     fn is_valid(&self, index: usize) -> VortexResult<bool> {
@@ -425,37 +416,6 @@ impl<V: VTable> Array for ArrayAdapter<V> {
         let scalar = <V::OperationsVTable as OperationsVTable<V>>::scalar_at(&self.0, index)?;
         assert_eq!(self.dtype(), scalar.dtype(), "Scalar dtype mismatch");
         Ok(scalar)
-    }
-
-    fn optimize(&self) -> VortexResult<ArrayRef> {
-        let result = <V::OperationsVTable as OperationsVTable<V>>::optimize(&self.0)?.into_array();
-
-        #[cfg(debug_assertions)]
-        {
-            let nbytes = self.0.nbytes();
-            let result_nbytes = result.nbytes();
-            assert!(
-                result_nbytes <= nbytes,
-                "optimize() made the array larger: {nbytes} bytes -> {result_nbytes} bytes",
-            );
-        }
-
-        assert_eq!(
-            self.dtype(),
-            result.dtype(),
-            "optimize() changed DType from {} to {}",
-            self.dtype(),
-            result.dtype()
-        );
-        assert_eq!(
-            result.len(),
-            self.len(),
-            "optimize() changed len from {} to {}",
-            self.len(),
-            result.len()
-        );
-
-        Ok(result)
     }
 
     fn is_valid(&self, index: usize) -> VortexResult<bool> {
